feat(cli): surface star prompt in `kesha status` and opt-in installs
Extracts the star-prompt logic from cli/install.ts into a shared
`maybeAskForStar(binPath, version, log)` in src/star.ts, then calls
it from `kesha status` (when the engine is installed) in addition to
the existing `kesha install` path.

The same on-disk marker (`<binPath>.star-seen`) gates both call
sites, so a user prompted by either path won't be re-prompted by
the other within the same major.minor. The version-bump gate
(`shouldShowStarPrompt`) is unchanged: first install + major-or-minor
bumps only, never on patch.

Why this surface: `kesha install --tts` / `--vad` opt-in flows
already routed through performInstall and so already prompted, but
users who opt in well after the first install are pinned to a
later seen-version and never see the prompt; folding the call into
`kesha status` gives a second non-spammy touchpoint where the user
has already opted into "tell me about this tool".
